+++ Webpages/PHP/deletePost.php
@@ -17,6 +17,10 @@
 
             $mysqli = new mysqli($servername, $username, $password, $dbname);
 
+            if ($mysqli->connect_error) {
+                die("Connection failed: " . $conn->connect_error);
+            }
+    
             $conn = $mysqli->prepare("DELETE FROM posts WHERE postID = ?");
 			$conn->bind_param('i', $postID);
 			$conn->execute();
